+++ backend/src/main/java/com/hvs/btctrader/bot/BotConfigRequest.java
@@ -74,15 +74,15 @@ public class BotConfigRequest {
 
 	@DecimalMin("0.5")
 	@DecimalMax("5.0")
-	private Double bbStdDev;
+	private BigDecimal bbStdDev;
 
 	@DecimalMin("0.001")
 	@DecimalMax("0.05")
-	private Double trendThreshold;
+	private BigDecimal trendThreshold;
 
 	@DecimalMin("0.01")
 	@DecimalMax("0.2")
-	private Double volatilityHigh;
+	private BigDecimal volatilityHigh;
 
 	@Min(10)
 	@Max(90)
@@ -228,27 +228,27 @@ public class BotConfigRequest {
 		this.bbPeriod = bbPeriod;
 	}
 
-	public Double getBbStdDev() {
+	public BigDecimal getBbStdDev() {
 		return bbStdDev;
 	}
 
-	public void setBbStdDev(Double bbStdDev) {
+	public void setBbStdDev(BigDecimal bbStdDev) {
 		this.bbStdDev = bbStdDev;
 	}
 
-	public Double getTrendThreshold() {
+	public BigDecimal getTrendThreshold() {
 		return trendThreshold;
 	}
 
-	public void setTrendThreshold(Double trendThreshold) {
+	public void setTrendThreshold(BigDecimal trendThreshold) {
 		this.trendThreshold = trendThreshold;
 	}
 
-	public Double getVolatilityHigh() {
+	public BigDecimal getVolatilityHigh() {
 		return volatilityHigh;
 	}
 
-	public void setVolatilityHigh(Double volatilityHigh) {
+	public void setVolatilityHigh(BigDecimal volatilityHigh) {
 		this.volatilityHigh = volatilityHigh;
 	}
